# Batch process join positions in JoinProbe