--- a/core/trino-main/src/main/java/io/trino/SystemSessionProperties.java
+++ b/core/trino-main/src/main/java/io/trino/SystemSessionProperties.java
@@ -143,6 +143,7 @@ public final class SystemSessionProperties
     public static final String MERGE_PROJECT_WITH_VALUES = "merge_project_with_values";
     public static final String TIME_ZONE_ID = "time_zone_id";
     public static final String LEGACY_CATALOG_ROLES = "legacy_catalog_roles";
+    public static final String VECTORIZED_JOIN_PROBE_ENABLED = "vectorized_join_probe_enabled";
 
     private final List<PropertyMetadata<?>> sessionProperties;
 
@@ -660,7 +661,11 @@ public final class SystemSessionProperties
                         LEGACY_CATALOG_ROLES,
                         "Enable legacy role management syntax that assumed all roles are catalog scoped",
                         featuresConfig.isLegacyCatalogRoles(),
-                        true));
+                        true),
+                booleanProperty(VECTORIZED_JOIN_PROBE_ENABLED,
+                        "Enable vectorized join probe processing",
+                        true,
+                        false));
     }
 
     @Override
@@ -1172,5 +1177,10 @@ public final class SystemSessionProperties
     public static boolean isLegacyCatalogRoles(Session session)
     {
         return session.getSystemProperty(LEGACY_CATALOG_ROLES, Boolean.class);
+    }
+
+    public static boolean isVectorizedJoinProbeEnabled(Session session)
+    {
+        return session.getSystemProperty(VECTORIZED_JOIN_PROBE_ENABLED, Boolean.class);
     }
 }