--- a/core/trino-main/src/main/java/io/trino/operator/join/JoinHash.java
+++ b/core/trino-main/src/main/java/io/trino/operator/join/JoinHash.java
@@ -156,7 +156,7 @@ public final class JoinHash
     @Override
     public boolean supportsCaching()
     {
-        return true;
+        return pagesHash.getPositionCount() > 1000;
     }
 
     @Override