… core/trino-main/src/main/java/io/trino/operator/join/LookupSource.java
… core/trino-main/src/main/java/io/trino/operator/join/PagesHash.java
… core/trino-main/src/main/java/io/trino/operator/join/PartitionedLookupSource.java
… core/trino-main/src/test/java/io/trino/operator/join/BenchmarkHashBuildAndJoinOperators.java
… core/trino-main/src/test/java/io/trino/operator/join/JoinTestUtils.java
… core/trino-main/src/test/java/io/trino/operator/join/TestHashJoinOperator.java
… core/trino-main/src/test/java/io/trino/operator/join/TestLookupJoinPageBuilder.java
… testing/trino-benchmark/src/main/java/io/trino/benchmark/HashBuildBenchmark.java
… testing/trino-benchmark/src/main/java/io/trino/benchmark/HashJoinBenchmark.java