Benchmarks before:
Benchmark                                             (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  Mode  Cnt    Score   Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint          0.1           bigint  avgt   30  217,846 ± 2,648  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint            1           bigint  avgt   30  206,986 ± 2,609  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint            2           bigint  avgt   30  125,916 ± 0,749  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint          0.1           bigint  avgt   30  215,422 ± 5,864  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint            1           bigint  avgt   30  418,456 ± 4,223  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint            2           bigint  avgt   30  250,294 ± 1,497  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.1           bigint  avgt   30  176,874 ± 1,664  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            1           bigint  avgt   30  202,235 ± 3,387  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            2           bigint  avgt   30  121,637 ± 0,928  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint          0.1           bigint  avgt   30  182,658 ± 2,173  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint            1           bigint  avgt   30  405,292 ± 6,822  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint            2           bigint  avgt   30  249,612 ± 2,348  ms/op

after:
Benchmark                                             (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  Mode  Cnt    Score   Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint          0.1           bigint  avgt   30   80,986 ± 3,045  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint            1           bigint  avgt   30   63,351 ± 0,384  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      1         bigint            2           bigint  avgt   30   61,030 ± 0,587  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint          0.1           bigint  avgt   30  103,966 ± 0,907  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint            1           bigint  avgt   30  190,340 ± 0,816  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash               false                      5         bigint            2           bigint  avgt   30  186,862 ± 1,864  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.1           bigint  avgt   30   80,280 ± 1,136  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            1           bigint  avgt   30   65,010 ± 0,655  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            2           bigint  avgt   30   60,309 ± 0,280  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint          0.1           bigint  avgt   30  102,268 ± 0,976  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint            1           bigint  avgt   30  190,286 ± 1,005  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      5         bigint            2           bigint  avgt   30  184,353 ± 1,160  ms/op