@sopel39 2021-09-20
```
Benchmark                                             (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  (partitionCount)  (vectorizationEnabled)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5           bigint                 1                    true  avgt   30  290,861 ± 19,458  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5           bigint                 1                   false  avgt   30  369,203 ± 26,782  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5           bigint                16                    true  avgt   30  336,473 ± 17,520  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5           bigint                16                   false  avgt   30  420,071 ± 32,089  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5              all                 1                    true  avgt   30  342,962 ± 17,455  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5              all                 1                   false  avgt   30  450,272 ± 34,538  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5              all                16                    true  avgt   30  408,102 ± 13,470  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint          0.5              all                16                   false  avgt   30  506,794 ± 26,384  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5           bigint                 1                    true  avgt   30  153,704 ±  4,470  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5           bigint                 1                   false  avgt   30  171,647 ±  4,498  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5           bigint                16                    true  avgt   30  221,681 ±  3,084  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5           bigint                16                   false  avgt   30  232,611 ± 10,945  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5              all                 1                    true  avgt   30  297,843 ± 14,642  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5              all                 1                   false  avgt   30  360,541 ± 36,997  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5              all                16                    true  avgt   30  393,449 ± 22,833  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                true                      1         bigint            5              all                16                   false  avgt   30  426,211 ± 20,278  ms/op

```